Add connection retry for dashboard
